// Capstone Project/src/app/pages/admin/admin-map-position/screen/types.ts
export enum adminMapPositonEnum {
  GET_MAP_POSITON_REQUEST = 'APP/ADMIN_PAGE/GET_MAP_POSITON_REQUEST',
  GET_MAP_POSITON_SUCCESS = 'APP/ADMIN_PAGE/GET_DMAP_POSITON_SUCCESS',
  GET_MAP_POSITON_FAILURE = 'APP/ADMIN_PAGE/GET_MAP_POSITON_FAILURE',

  CREATE_MAP_POSITON_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_MAP_POSITON_REQUEST',
  CREATE_MAP_POSITON_SUCCESS = 'APP/ADMIN_PAGE/CREATE_MAP_POSITON_SUCCESS',
  CREATE_MAP_POSITON_FAILURE = 'APP/ADMIN_PAGE/CREATE_MAP_POSITON_FAILURE',

  UPDATE_MAP_POSITON_REQUEST = 'APP/ADMIN_PAGE/UPDATE_MAP_POSITON_REQUEST',
  UPDATE_MAP_POSITON_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_MAP_POSITON_SUCCESS',
  UPDATE_MAP_POSITON_FAILURE = 'APP/ADMIN_PAGE/UPDATE_MAP_POSITON_FAILURE',

  DELETE_MAP_POSITON_REQUEST = 'APP/ADMIN_PAGE/DELETE_MAP_POSITON_REQUEST',
  DELETE_MAP_POSITON_SUCCESS = 'APP/ADMIN_PAGE/DELETE_MAP_POSITON_SUCCESS',
  DELETE_MAP_POSITON_FAILURE = 'APP/ADMIN_PAGE/DELETE_MAP_POSITON_FAILURE',

  CLEAR_MAP_POSITON_STATE = 'APP/ADMIN_PAGE/CLEAR_MAP_POSITON_STATE',
}